+++ src/main/java/wooteco/subway/application/SectionService.java
@@ -1,6 +1,5 @@
 package wooteco.subway.application;
 
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.SectionDao;
@@ -8,6 +7,8 @@ import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.NoSuchStationException;
+
+import java.util.List;
 
 @Service
 @Transactional
