+++ src/main/java/wooteco/subway/application/StationService.java
@@ -1,7 +1,5 @@
 package wooteco.subway.application;
 
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.StationDao;
@@ -9,6 +7,9 @@ import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
 import wooteco.subway.exception.DuplicateStationNameException;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
