+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,9 +1,10 @@
 package wooteco.subway.dao;
 
-import java.util.List;
-import java.util.Optional;
 import wooteco.subway.domain.Line;
 import wooteco.subway.exception.NoSuchLineException;
+
+import java.util.List;
+import java.util.Optional;
 
 public interface LineDao {
     Line save(Line line) throws IllegalArgumentException;
