+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -1,7 +1,8 @@
 package wooteco.subway.dao;
 
-import java.util.List;
 import wooteco.subway.domain.Section;
+
+import java.util.List;
 
 public interface SectionDao {
     Section save(Long lineId, Section section);
