+++ src/main/java/wooteco/subway/dto/PathResponse.java
@@ -1,8 +1,9 @@
 package wooteco.subway.dto;
 
+import wooteco.subway.domain.Station;
+
 import java.util.List;
 import java.util.stream.Collectors;
-import wooteco.subway.domain.Station;
 
 public class PathResponse {
 
@@ -10,7 +11,8 @@ public class PathResponse {
     private int distance;
     private int fare;
 
-    private PathResponse(){}
+    private PathResponse() {
+    }
 
     public static PathResponse from(final List<Station> stations, final int distance, final int fare) {
         List<StationResponse> stationResponses = stations.stream()
