+++ src/main/java/wooteco/subway/dto/StationResponse.java
@@ -6,7 +6,8 @@ public class StationResponse {
     private Long id;
     private String name;
 
-    private StationResponse(){}
+    private StationResponse() {
+    }
 
     public StationResponse(Long id, String name) {
         this.id = id;
