+++ src/main/java/wooteco/subway/presentation/PathController.java
@@ -1,7 +1,10 @@
 package wooteco.subway.presentation;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.application.PathService;
 import wooteco.subway.dto.PathRequest;
 import wooteco.subway.dto.PathResponse;
