+++ src/main/java/wooteco/subway/presentation/StationController.java
@@ -1,18 +1,13 @@
 package wooteco.subway.presentation;
 
-import java.net.URI;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import wooteco.subway.application.StationService;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
+
+import java.net.URI;
+import java.util.List;
 
 @RestController
 @RequestMapping("/stations")
